Update create_newcase to be ice sheet model agnostic
Previously, create_newcase used a variable CISM_GRID to identify what
grid CISM will be using internally (which is different than the GLC grid
which is the same as the LND grid).  I've changed CISM_GRID to
ICESHEET_GRID everywhere to make it clear that this is the grid used by
the chosen ice sheet model, and not for CISM specifically (which is no
longer then only ISM).